Move Extractor calls into a single thread
Moves all calls to the Extractor library into a single thread that's
separate from both the main thread and the user interface thread.

This is achieved using a thread pool with just a single thread.
Restricting the pool to one thread is important as calls to the
Extractor library don't appear to be threadsafe.