Allow multiple comment models
Removes the dependence on a couple of comment models with global scope
and switches to using models at page scope.

Improves lifetime handling so that if the models are destroyed before
the Extractor futures resolve, the model isn't used, thereby avoiding
invalid memeory reads or writes.

Applies the same technique to the MediaInfo object so that a hash table
is no longer needed to store values semi-persistently.