Avoid searching on empty search string
In case the search field is cleared an empty search string is used. If
sent to the server this triggers a "NothingFoundException" exception and
returns no results. The round trip and stacktrace can both be avoided by
checking for empty search strings in advance and simply clearing the
search results model rather than triggering a search.

This change applies the client-side logic for checking empty strings and
clearing the model without triggering an actual search.